常用模块列子/requirements.txt
aenum==2.2.3
aes==1.0.0
aiofiles==0.4.0
aiohttp==3.6.2
amqp==2.5.2
appnope==0.1.0
APScheduler==3.6.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.1.0
cacheout==0.11.2
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cos-python-sdk-v5==1.7.6
cypari==2.3.0
decorator==4.4.1
dicttoxml==1.7.4
Django==3.0
fb==0.4.0
fbchat==1.9.4
Flask==1.1.1
furl==2.1.0
future==0.18.2
FXrays==1.3.3
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.1.0
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
jieba==0.40
Jinja2==2.10.3
kafka==1.3.5
kafka-python==1.4.7
kombu==4.6.6
MarkupSafe==1.1.1
more-itertools==8.0.0
msgpack==0.6.2
multidict==4.6.1
munch==2.5.0
mysql-connector==2.2.9
Naked==0.1.31
networkx==2.4
orderedmultidict==1.0.1
packaging==19.2
paho-mqtt==1.5.0
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.1
plink==2.3
pluggy==0.13.1
prompt-toolkit==3.0.3
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.0
pycrypto==2.6.1
Pygments==2.5.2
pynsq==0.8.3
pyparsing==2.4.5
pypinyin==0.36.0
pypng==0.0.20
pytest==5.3.1
python-snappy==0.5.4
pytz==2019.3
PyYAML==5.3
qcloud-cos-py3==0.1.4
redis==3.3.11
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
shellescape==3.4.1
singledispatch==3.4.0.3
six==1.13.0
soupsieve==1.9.5
spherogram==1.8.2
SQLAlchemy==1.3.11
sqlparse==0.3.0
tornado==4.5.3
tqdm==4.39.0
traitlets==4.3.3
tzlocal==2.0.0
ujson==1.35
urllib3==1.25.7
uvloop==0.14.0
vine==1.3.0
wcwidth==0.1.7
websockets==8.1
Werkzeug==0.16.0
xmltodict==0.12.0
yarl==1.4.1
zipp==0.6.0